YAY-14 refactor: move FilterModal to separate component
